feat: move Topic and campaign to a separate file and removed all import *
