+++ src/furet/app/windows/raaDetailsWindow.py
@@ -1,3 +1,4 @@
+from furet.app.utils import DECREE_COLUMNS
 from furet.app.widgets.objectTableModel import ObjectTableColumn
 from furet.app.widgets.objectTableWidget import ObjectTableWidget
 from furet.app.widgets.optionalDateEdit import NONE_DATE
@@ -26,36 +27,29 @@ class RaaDetailsWindow(QtWidgets.QDialog):
 
         self._separator = TextSeparatorWidget("Arrêtés")
         self._layout.addWidget(self._separator)
-        self._decreeLabel = TextSeparatorWidget(f"{len(decrees)} arrêté(s) pertinent(s) sur {raa.decreeCount}")
-        self._layout.addWidget(self._separator)
 
-        self._decreesTable = ObjectTableWidget(self._decrees, [
-            ObjectTableColumn("Campagnes", lambda v: v.campaigns, lambda v: ", ".join(map(str, v))),
-            ObjectTableColumn("Sujets", lambda v: v.topics, lambda v: ", ".join(map(str, v))),
-            ObjectTableColumn("Titre", lambda v: v.title),
-            ObjectTableColumn("À compléter", lambda v: v.missingValues(False), lambda v: f"{v} champs" if v else ""), # TODO label not visible
-        ])
-        self._decreesTable.doubleClicked.connect(self.onDblClickTableRow)
+        self._decreesTable = ObjectTableWidget(self._decrees, DECREE_COLUMNS[3:7])
+        self._decreesTable.doubleClicked.connect(lambda i: self.showDecreeDetailsWindow(self._decreesTable.itemAt(i.row())))
         self._layout.addWidget(self._decreesTable)
+
         self._buttons = QtWidgets.QDialogButtonBox(standardButtons=QtWidgets.QDialogButtonBox.StandardButton.Save | QtWidgets.QDialogButtonBox.StandardButton.Close)
-        self._buttons.accepted.connect(self.save)
+        self._buttons.accepted.connect(self.accept)
         self._buttons.rejected.connect(self.reject)
         self._layout.addWidget(self._buttons)
 
-    def onDblClickTableRow(self, index: QtCore.QModelIndex):
-        decree = self._decreesTable.itemAt(index.row())
-        window, created = windowManager.showWindow(DecreeDetailsWindow, decree.id, args=(decree,), kwargs={ "noRaa":True}) # TODO handle RAA info edit / disable it
-        window.accepted.connect(self.onDecreeSaved, type=QtCore.Qt.ConnectionType.UniqueConnection)
+    def showDecreeDetailsWindow(self, decree: Decree):
+        window, created = windowManager.showWindow(DecreeDetailsWindow, decree.id, args=(decree,), kwargs={ "noRaa":True })
+        window.accepted.connect(self.updateDecrees, type=QtCore.Qt.ConnectionType.UniqueConnection)
 
-    def save(self):
+    def accept(self) -> None:
         raa = self._raaWidget.raa()
         repository.updateRaa(raa.id, raa)
 
         # TODO find a way to remove this, currently, the decree does not know that tha raa was changed and fetching returns an old versions as the decrees file is not changed
         for d in self._decrees:
             repository.updateDecree(d.id, d)
-        self.accept()
-    
-    def onDecreeSaved(self):
+        super().accept()
+
+    def updateDecrees(self):
         self._decreesTable.setItems([repository.getDecreeById(d.id) for d in self._decrees])
