+++ src/furet/processing/separation.py
@@ -1,5 +1,7 @@
 from furet import repository
-from furet.types.decree import *
+from furet.types.decree import Decree
+from furet.types.raa import RAA
+
 from furet.processing.getPdfText import extractText
 from datetime import date
