+++ src/furet/repository/__init__.py
@@ -1,10 +1,13 @@
+from datetime import date
 import os
 from furet import settings
 from furet.repository import csvdb
-from furet.types.department import *
-from furet.types.decree import *
+from furet.types.department import Department
+from furet.types.decree import DocumentType, Decree
+from furet.types.campaign import Campaign, Topic
+from furet.types.raa import RAA
 from PySide6 import QtCore
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 
 def setup():
     settings.setDefaultValue("repository.csv-root", os.path.join(QtCore.QStandardPaths.writableLocation(QtCore.QStandardPaths.StandardLocation.AppDataLocation), 'database'))
